src/backend/requirements.txt
aiofiles
azure-ai-documentintelligence>=1.0.0
azure-ai-inference>=1.0.0b6
azure-identity
tqdm 
python-dotenv
requests
openai>=1.50.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pillow
rich
azure-storage-blob>=12.24.0
--extra-index-url https://pkgs.dev.azure.com/azure-sdk/public/_packaging/azure-sdk-for-python/pypi/simple/
azure-search-documents==11.6.0a20250505003
pyjwt
fastapi
uvicorn